+++ sdk/contracts/core/wesdt-swap/tests/wegld_swap_scenario_rs_test.rs
@@ -2,10 +2,10 @@ use multiversx_sc_scenario::*;
 
 fn world() -> ScenarioWorld {
     let mut blockchain = ScenarioWorld::new();
-    blockchain.set_current_dir_from_workspace("contracts/core/wegld-swap");
+    blockchain.set_current_dir_from_workspace("contracts/core/wesdt-swap");
 
     blockchain.register_contract(
-        "file:output/multiversx-wegld-swap-sc.wasm",
+        "file:output/multiversx-wesdt-swap-sc.wasm",
         multiversx_wegld_swap_sc::ContractBuilder,
     );
     blockchain